Refactor on events builder folder (#199)
* Refactor

* Code analysis

* Fix

* Fix
