+++ app/services/builders/chartkick/project_data.rb
@@ -2,7 +2,7 @@ module Builders
   module Chartkick
     class ProjectData < Builders::Chartkick::Base
       def call
-        project = Project.find(@entity_id)
+        project = ::Project.find(@entity_id)
         metrics = project.metrics.where(@query)
         [{ name: project.name, data: build_data(metrics) }]
       end
